avcodec/movtextenc: fix writing to bytestream on BE arches
Fixes fate-binsub-movtextenc on PPC64

Currently tags are written in reverse order on BE arches. This is fixed
by using MKBETAG() and AV_32() to be arch agnostics.

Also s->font_count is of type int. On BE arches with 32bits ints,
count = AV_RB16(&s->font_count) will read two most significant bytes
instead of the least signifcant bytes. This is fixed by assiging
s->font_count to count first.

The final change is modyfying the type of len. On BE arches
the most signifanct byte of the int was written instead of the least
significant byte.

Signed-off-by: Andriy Gelman <andriy.gelman@gmail.com>